kernel: timeout: Extract shared fucntions into common files
dlist backend and the min-heap backend need access to several
functions that are common.

* Add kernel/include/timeout_priv.h with extern declarations for the
  three shared variables and a static inline elapsed()
* make curr_tick, timeout_lock, and announce_remaining non-static in
  timeout.c
* Move local elapsed() definition from timeout.c to timeout_priv.h
* Fucntions that are not specific to either the dlist or the
  heap backend are moved to a new timeout_common.c to avoid
  duplication.

  sys_clock_tick_get, sys_clock_tick_get_32, z_impl_k_uptime_ticks,
  z_vrfy_k_uptime_ticks, sys_timepoint_calc, sys_timepoint_timeout,
  sys_clock_lock, sys_clock_unlock, z_impl_sys_clock_tick_set,
  z_vrfy_sys_clock_tick_set.

  No functional change.
