+++ kernel/timeout.c
@@ -12,10 +12,11 @@
 #include <zephyr/drivers/timer/system_timer.h>
 #include <zephyr/sys_clock.h>
 #include <zephyr/llext/symbol.h>
+#include "timeout_priv.h"
 
 #include <timeslicing.h>
 
-static uint64_t curr_tick;
+uint64_t curr_tick;
 
 static sys_dlist_t timeout_list = SYS_DLIST_STATIC_INIT(&timeout_list);
 
@@ -23,10 +24,10 @@ static sys_dlist_t timeout_list = SYS_DLIST_STATIC_INIT(&timeout_list);
  * The timeout code shall take no locks other than its own (timeout_lock), nor
  * shall it call any other subsystem while holding this lock.
  */
-static struct k_spinlock timeout_lock;
+struct k_spinlock timeout_lock;
 
 /* Ticks left to process in the currently-executing sys_clock_announce() */
-static int announce_remaining;
+int announce_remaining;
 
 static struct _timeout *first(void)
 {
@@ -49,27 +50,6 @@ static void remove_timeout(struct _timeout *t)
 	}
 
 	sys_dlist_remove(&t->node);
-}
-
-static int32_t elapsed(void)
-{
-	/* While sys_clock_announce() is executing, new relative timeouts will be
-	 * scheduled relatively to the currently firing timeout's original tick
-	 * value (=curr_tick) rather than relative to the current
-	 * sys_clock_elapsed().
-	 *
-	 * This means that timeouts being scheduled from within timeout callbacks
-	 * will be scheduled at well-defined offsets from the currently firing
-	 * timeout.
-	 *
-	 * As a side effect, the same will happen if an ISR with higher priority
-	 * preempts a timeout callback and schedules a timeout.
-	 *
-	 * The distinction is implemented by looking at announce_remaining which
-	 * will be non-zero while sys_clock_announce() is executing and zero
-	 * otherwise.
-	 */
-	return announce_remaining == 0 ? sys_clock_elapsed() : 0U;
 }
 
 static int32_t next_timeout(int32_t ticks_elapsed)
@@ -295,103 +275,3 @@ void sys_clock_announce_locked(int32_t ticks, k_spinlock_key_t key)
 	z_time_slice();
 #endif /* CONFIG_TIMESLICING */
 }
-
-#if defined(CONFIG_SMP) || defined(CONFIG_SPIN_VALIDATE)
-k_spinlock_key_t sys_clock_lock(void)
-{
-	return k_spin_lock(&timeout_lock);
-}
-
-void sys_clock_unlock(k_spinlock_key_t key)
-{
-	k_spin_unlock(&timeout_lock, key);
-}
-#endif
-
-#if defined(CONFIG_TEST) || defined(CONFIG_ASSERT)
-bool sys_clock_is_locked(void)
-{
-	return z_spin_is_locked(&timeout_lock);
-}
-#endif
-
-int64_t sys_clock_tick_get(void)
-{
-	uint64_t t = 0U;
-
-	K_SPINLOCK(&timeout_lock) {
-		t = curr_tick + elapsed();
-	}
-	return t;
-}
-
-uint32_t sys_clock_tick_get_32(void)
-{
-#ifdef CONFIG_TICKLESS_KERNEL
-	return (uint32_t)sys_clock_tick_get();
-#else
-	return (uint32_t)curr_tick;
-#endif /* CONFIG_TICKLESS_KERNEL */
-}
-
-int64_t z_impl_k_uptime_ticks(void)
-{
-	return sys_clock_tick_get();
-}
-
-#ifdef CONFIG_USERSPACE
-static inline int64_t z_vrfy_k_uptime_ticks(void)
-{
-	return z_impl_k_uptime_ticks();
-}
-#include <zephyr/syscalls/k_uptime_ticks_mrsh.c>
-#endif /* CONFIG_USERSPACE */
-
-k_timepoint_t sys_timepoint_calc(k_timeout_t timeout)
-{
-	k_timepoint_t timepoint;
-
-	if (K_TIMEOUT_EQ(timeout, K_FOREVER)) {
-		timepoint.tick = UINT64_MAX;
-	} else if (K_TIMEOUT_EQ(timeout, K_NO_WAIT)) {
-		timepoint.tick = 0;
-	} else {
-		k_ticks_t dt = timeout.ticks;
-
-		if (Z_IS_TIMEOUT_RELATIVE(timeout)) {
-			timepoint.tick = sys_clock_tick_get() + max(1, dt);
-		} else {
-			timepoint.tick = Z_TICK_ABS(dt);
-		}
-	}
-
-	return timepoint;
-}
-
-k_timeout_t sys_timepoint_timeout(k_timepoint_t timepoint)
-{
-	uint64_t now, remaining;
-
-	if (timepoint.tick == UINT64_MAX) {
-		return K_FOREVER;
-	}
-	if (timepoint.tick == 0) {
-		return K_NO_WAIT;
-	}
-
-	now = sys_clock_tick_get();
-	remaining = (timepoint.tick > now) ? (timepoint.tick - now) : 0;
-	return K_TICKS(remaining);
-}
-
-#ifdef CONFIG_ZTEST
-void z_impl_sys_clock_tick_set(uint64_t tick)
-{
-	curr_tick = tick;
-}
-
-void z_vrfy_sys_clock_tick_set(uint64_t tick)
-{
-	z_impl_sys_clock_tick_set(tick);
-}
-#endif /* CONFIG_ZTEST */
