+++ kernel/timeout_common.c
@@ -1,0 +1,119 @@
+/**
+ * Copyright (c) 2018 Intel Corporation
+ * Copyright (c) 2026 Aerlync Labs Inc.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/*
+ * Kernel timeout subsystem : backend-independent utilities.
+ *
+ */
+
+#include <zephyr/kernel.h>
+#include <zephyr/internal/syscall_handler.h>
+#include <zephyr/sys_clock.h>
+#include "timeout_priv.h"
+
+int64_t z_impl_k_uptime_ticks(void)
+{
+	return sys_clock_tick_get();
+}
+
+#ifdef CONFIG_USERSPACE
+static inline int64_t z_vrfy_k_uptime_ticks(void)
+{
+	return z_impl_k_uptime_ticks();
+}
+#include <zephyr/syscalls/k_uptime_ticks_mrsh.c>
+#endif /* CONFIG_USERSPACE */
+
+k_timepoint_t sys_timepoint_calc(k_timeout_t timeout)
+{
+	k_timepoint_t timepoint;
+
+	if (K_TIMEOUT_EQ(timeout, K_FOREVER)) {
+		timepoint.tick = UINT64_MAX;
+	} else if (K_TIMEOUT_EQ(timeout, K_NO_WAIT)) {
+		timepoint.tick = 0;
+	} else {
+		k_ticks_t dt = timeout.ticks;
+
+		if (Z_IS_TIMEOUT_RELATIVE(timeout)) {
+			timepoint.tick = sys_clock_tick_get() + MAX(1, dt);
+		} else {
+			timepoint.tick = Z_TICK_ABS(dt);
+		}
+	}
+
+	return timepoint;
+}
+
+k_timeout_t sys_timepoint_timeout(k_timepoint_t timepoint)
+{
+	uint64_t now, remaining;
+
+	if (timepoint.tick == UINT64_MAX) {
+		return K_FOREVER;
+	}
+
+	if (timepoint.tick == 0) {
+		return K_NO_WAIT;
+	}
+
+	now = sys_clock_tick_get();
+	remaining = (timepoint.tick > now) ? (timepoint.tick - now) : 0;
+
+	return K_TICKS(remaining);
+}
+
+int64_t sys_clock_tick_get(void)
+{
+	uint64_t t = 0U;
+
+	K_SPINLOCK(&timeout_lock) {
+		t = curr_tick + elapsed();
+	}
+
+	return t;
+}
+
+uint32_t sys_clock_tick_get_32(void)
+{
+#ifdef CONFIG_TICKLESS_KERNEL
+	return (uint32_t)sys_clock_tick_get();
+#else
+	return (uint32_t)curr_tick;
+#endif /* CONFIG_TICKLESS_KERNEL */
+}
+
+#if defined(CONFIG_SMP) || defined(CONFIG_SPIN_VALIDATE)
+k_spinlock_key_t sys_clock_lock(void)
+{
+	return k_spin_lock(&timeout_lock);
+}
+
+void sys_clock_unlock(k_spinlock_key_t key)
+{
+	k_spin_unlock(&timeout_lock, key);
+}
+#endif
+
+#if defined(CONFIG_TEST) || defined(CONFIG_ASSERT)
+bool sys_clock_is_locked(void)
+{
+	return z_spin_is_locked(&timeout_lock);
+}
+#endif
+
+#ifdef CONFIG_ZTEST
+void z_impl_sys_clock_tick_set(uint64_t tick)
+{
+	curr_tick = tick;
+}
+
+void z_vrfy_sys_clock_tick_set(uint64_t tick)
+{
+	z_impl_sys_clock_tick_set(tick);
+}
+#endif /* CONFIG_ZTEST */
